backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
